third_party/blink/web_tests/platform/generic/http/tests/inspector-protocol/fenced-frame/reload-fenced-frame.https-expected.txt

Tests that Page.reload() in a fenced frame is not allowed.
Page.reload() from a fenced frame:
PASS: Reload is only supported for top-level frames
